+++ src/java/org/smoothbuild/task/exec/SmoothExecutor.java
@@ -7,6 +7,7 @@ import org.smoothbuild.lang.function.base.Function;
 import org.smoothbuild.lang.function.base.Module;
 import org.smoothbuild.lang.function.base.Name;
 import org.smoothbuild.message.listen.ErrorMessageException;
+import org.smoothbuild.task.exec.err.NoFunctionSpecifiedError;
 import org.smoothbuild.task.exec.err.UnknownFunctionError;
 
 public class SmoothExecutor {
@@ -21,6 +22,9 @@ public class SmoothExecutor {
     CommandLineArguments args = executionData.args();
     Module module = executionData.module();
 
+    if (args.functionsToRun().isEmpty()) {
+      throw new ErrorMessageException(new NoFunctionSpecifiedError(module.availableNames()));
+    }
     for (Name name : args.functionsToRun()) {
       Function function = module.getFunction(name);
       if (function == null) {